Adding some page fault handling and fixed a bug in screen scrolling by chaning mem_cpy. Need to implement frame handling before paging is really done.. also need to not just hardcode one page lol

+++ drivers/screen.c
@@ -74,7 +74,7 @@ void print_backspace() {
  */
 int scroll_screen() {
     // shift data up by one row
-    mem_cpy((uint32_t*)VIDEO_ADDRESS,
+    mem_cpy((uint8_t*)VIDEO_ADDRESS,
      (uint8_t*)(VIDEO_ADDRESS + (2 * MAX_COLUMNS)),
       MAX_COLUMNS * MAX_ROWS * 2);
     // blank out the bottom row
