+++ libc/mem.h
@@ -8,7 +8,7 @@
  */
 extern uint32_t placement_addr;
 
-void mem_cpy(uint32_t *to_address, uint8_t *from_address, int size);
+void mem_cpy(uint8_t *to_address, uint8_t *from_address, int size);
 void mem_set(uint32_t* dest, uint8_t val, uint32_t size);
 uint32_t mallok_internal(size_t size, int align, uint32_t *phys_addr);
 uint32_t mallok_a(uint32_t size);
